refactor: move asset result generation
Move asset result out of graph result. It makes more sense to use
GraphResult to generate/map to asset result. It makes it more clear
where asset result is used.
