+++ src/libecalc/presentation/json_result/aggregators.py
